Added AAP server process.
Added AAP server process that starts up and heart beat the applications
presence to the Auth Server and connections a channel connection with a
net.Conn to the target server.

Refactored dynamic labels out from the regular SSH server and into its
own package so it can be re-used by both the SSH server as well as the
application server.

Updated heart beat code to support heart beats for different kinds of
services.Server (SSH as well as application).