Added JWT "CA" and API to Auth Server.
Added a JWT "CA" to the Auth Server. The JWT CA is not a real CA in that
it does not sign certificates, but it behaves like one in terms of
storage, access, and rotation.

Added GenerateJWT GRPC API to the Auth Server. Used by AAP web
server to request signed JWT tokens which will be embedded within the
request headers and sent to the target application.

Added /webapi/certs endpoint to proxy web server to allow fetching of
JWT public keys which can be used to verify the JWT token by the target
application.